# src/backend/CompanyHero.Api/Contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CompanyHero API",
    "description": "Verbindlicher Vertrag nach A-009 und K13: Kennungen und Dezimalwerte als Strings, Zeitpunkte mit Offset, Null und fehlendes Feld unterscheidbar.",
    "version": "1"
  },
  "paths": {
...
    "/api/branding/theme": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "GetTheme",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "PublishTheme",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/branding/theme/versions/{version}": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "GetThemeVersion",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/branding/theme/preview": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "PreviewTheme",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemePreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/branding/tenants/{tenantId}/manifest.webmanifest": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "GetTenantManifest",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/manifest+json": {
                "schema": {
                  "$ref": "#/components/schemas/WebManifest"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/branding/tenants/{tenantId}/icons/{name}": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "GetTenantIcon",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnredeDto": {
        "enum": [
          "du",
          "sie"
        ]
      },
...
      "BezeichnungenDto": {
        "required": [
          "punkte",
          "serie",
          "stufen"
        ],
        "type": "object",
        "properties": {
          "punkte": {
            "type": "string"
          },
          "serie": {
            "type": "string"
          },
          "stufen": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ErsetzungDto": {
        "required": [
          "modus",
          "rolle",
          "angefordert",
          "ersetzt",
          "grund"
        ],
        "type": "object",
        "properties": {
          "modus": {
            "$ref": "#/components/schemas/ThemeModeDto"
          },
          "rolle": {
            "type": "string"
          },
          "angefordert": {
            "type": "string"
          },
          "ersetzt": {
            "type": "string"
          },
          "grund": {
            "type": "string"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "required": [
          "errors"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "KontrastDto": {
        "required": [
          "modus",
          "vordergrund",
          "hintergrund",
          "gefordert",
          "erreicht"
        ],
        "type": "object",
        "properties": {
          "modus": {
            "$ref": "#/components/schemas/ThemeModeDto"
          },
          "vordergrund": {
            "type": "string"
          },
          "hintergrund": {
            "type": "string"
          },
          "gefordert": {
            "type": "string"
          },
          "erreicht": {
            "type": "string"
          }
        }
      },
...
      "ManifestIcon": {
        "required": [
          "src",
          "sizes",
          "type",
          "purpose"
        ],
        "type": "object",
        "properties": {
          "src": {
            "type": "string"
          },
          "sizes": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "purpose": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ThemeDocumentRequest": {
        "required": [
          "schema",
          "produktname",
          "saatfarbe",
          "akzent",
          "anrede",
          "tonalitaet",
          "bezeichnungen",
          "willkommenstext"
        ],
        "type": "object",
        "properties": {
          "schema": {
            "type": "integer",
            "format": "int32"
          },
          "produktname": {
            "type": "string"
          },
          "saatfarbe": {
            "type": "string"
          },
          "akzent": {
            "type": [
              "null",
              "string"
            ]
          },
          "anrede": {
            "$ref": "#/components/schemas/AnredeDto"
          },
          "tonalitaet": {
            "$ref": "#/components/schemas/TonalitaetDto"
          },
          "bezeichnungen": {
            "$ref": "#/components/schemas/BezeichnungenDto"
          },
          "willkommenstext": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ThemeModeDto": {
        "enum": [
          "hell",
          "dunkel"
        ]
      },
      "ThemePreviewResponse": {
        "required": [
          "verfahren",
          "hell",
          "dunkel",
          "ersetzungen",
          "kontrast"
        ],
        "type": "object",
        "properties": {
          "verfahren": {
            "type": "string"
          },
          "hell": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "dunkel": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "ersetzungen": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErsetzungDto"
            }
          },
          "kontrast": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KontrastDto"
            }
          }
        }
      },
      "ThemeResponse": {
        "required": [
          "tenantId",
          "version",
          "verfahren",
          "schema",
          "produktname",
          "anrede",
          "tonalitaet",
          "bezeichnungen",
          "willkommenstext",
          "hell",
          "dunkel",
          "ersetzungen",
          "texte",
          "manifestPfad",
          "veroeffentlichtAm"
        ],
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "verfahren": {
            "type": "string"
          },
          "schema": {
            "type": "integer",
            "format": "int32"
          },
          "produktname": {
            "type": "string"
          },
          "anrede": {
            "$ref": "#/components/schemas/AnredeDto"
          },
          "tonalitaet": {
            "$ref": "#/components/schemas/TonalitaetDto"
          },
          "bezeichnungen": {
            "$ref": "#/components/schemas/BezeichnungenDto"
          },
          "willkommenstext": {
            "type": [
              "null",
              "string"
            ]
          },
          "hell": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "dunkel": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "ersetzungen": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErsetzungDto"
            }
          },
          "texte": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "manifestPfad": {
            "type": "string"
          },
          "veroeffentlichtAm": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "TonalitaetDto": {
        "enum": [
          "sachlich",
          "freundlich",
          "motivierend"
        ]
      },
...
      "WebManifest": {
        "required": [
          "id",
          "name",
          "short_name",
          "start_url",
          "scope",
          "display",
          "lang",
          "theme_color",
          "background_color",
          "icons",
          "orientation"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "short_name": {
            "type": "string"
          },
          "start_url": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "display": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "theme_color": {
            "type": "string"
          },
          "background_color": {
            "type": "string"
          },
          "icons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManifestIcon"
            }
          },
          "orientation": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}